Add definitions to the environment. Remove hack in `ast`.
`Env` now maps `Id`s to `Binding`s, which contain the definitions of the
symbols as well as their types.

I don't know if this is strictly necessary for pure type systems, but I want to
use the same parser for all of them, and the easiest way to do it is to use
names, rather than very special symbols like ∗ and ◻, to refer to `Type` and
`Kind`. The axioms then need to be stored in the environment, and this change
accomplishes that.

This *is* necessary to do anything interesting with dependent types. Giving the
sorts regular names and having them participate in inference seems inevitable;
but more importantly a typecheck-time environment has to be able to access
definitions in order to determine whether or not `f 0` is definitionally
equivalent to `1`, without which unification is very weak.